SNAPSHOT Fix ScoreTest assertion

+++ modules/core-test/src/main/kotlin/org/coner/trailer/ScoreAssertk.kt
@@ -4,6 +4,7 @@ import assertk.Assert
 import assertk.assertions.isEqualTo
 import assertk.assertions.prop
 import org.coner.trailer.eventresults.Score
+import java.math.BigDecimal
 
 fun Assert<Score>.value() = prop("value") { it.value }
-fun Assert<Score>.hasValue(expected: Int) = value().isEqualTo(expected)
+fun Assert<Score>.hasValue(expected: BigDecimal) = value().isEqualTo(expected)
